Move node into a separate file
